Upgraded to .Net 6

+++ Mvc.RoleAuthorization/Services/DataAccessService.cs
@@ -1,70 +1,83 @@
 ﻿using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Caching.Memory;
 using Mvc.RoleAuthorization.Data;
 using Mvc.RoleAuthorization.Models;
-using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Security.Claims;
-using System.Threading.Tasks;
 
 namespace Mvc.RoleAuthorization.Services
 {
 	public class DataAccessService : IDataAccessService
 	{
+		private readonly IMemoryCache _cache;
 		private readonly ApplicationDbContext _context;
 
-		public DataAccessService(ApplicationDbContext context)
+		public DataAccessService(ApplicationDbContext context, IMemoryCache cache)
 		{
-			_context = context;
+			_cache = cache ?? throw new ArgumentNullException(nameof(cache));
+			_context = context ?? throw new ArgumentNullException(nameof(context));
 		}
 
 		public async Task<List<NavigationMenuViewModel>> GetMenuItemsAsync(ClaimsPrincipal principal)
 		{
-			var isAuthenticated = principal.Identity.IsAuthenticated;
-			if (!isAuthenticated)
+			var isAuthenticated = principal.Identity?.IsAuthenticated;
+			if (isAuthenticated == false)
+			{
 				return new List<NavigationMenuViewModel>();
+			}
 
 			var roleIds = await GetUserRoleIds(principal);
-			var data = await (from menu in _context.RoleMenuPermission
-							  where roleIds.Contains(menu.RoleId)
-							  select menu)
+
+			var permissions = await _cache.GetOrCreateAsync("Permissions",
+				async x => await (from menu in _context.NavigationMenu select menu).ToListAsync());
+
+			var rolePermissions = await _cache.GetOrCreateAsync("RolePermissions",
+				async x => await (from menu in _context.RoleMenuPermission select menu).Include(x => x.NavigationMenu).ToListAsync());
+
+			var data = (from menu in rolePermissions
+						join p in permissions on menu.NavigationMenuId equals p.Id
+						where roleIds.Contains(menu.RoleId)
+						select p)
 							  .Select(m => new NavigationMenuViewModel()
 							  {
-								  Id = m.NavigationMenu.Id,
-								  Name = m.NavigationMenu.Name,
-								  Area = m.NavigationMenu.Area,
-								  ActionName = m.NavigationMenu.ActionName,
-								  ControllerName = m.NavigationMenu.ControllerName,
-								  IsExternal = m.NavigationMenu.IsExternal,
-								  ExternalUrl = m.NavigationMenu.ExternalUrl,
-								  DisplayOrder = m.NavigationMenu.DisplayOrder,
-								  ParentMenuId = m.NavigationMenu.ParentMenuId,
-								  Visible = m.NavigationMenu.Visible,
-							  }).Distinct().ToListAsync();
+								  Id = m.Id,
+								  Name = m.Name,
+								  Area = m.Area,
+								  Visible = m.Visible,
+								  IsExternal = m.IsExternal,
+								  ActionName = m.ActionName,
+								  ExternalUrl = m.ExternalUrl,
+								  DisplayOrder = m.DisplayOrder,
+								  ParentMenuId = m.ParentMenuId,
+								  ControllerName = m.ControllerName,
+							  }).Distinct().ToList();
 
 			return data;
 		}
 
-		public async Task<bool> GetMenuItemsAsync(ClaimsPrincipal ctx, string ctrl, string act)
+		public async Task<bool> GetMenuItemsAsync(ClaimsPrincipal ctx, string? ctrl, string? act)
 		{
 			var result = false;
 			var roleIds = await GetUserRoleIds(ctx);
 			var data = await (from menu in _context.RoleMenuPermission
 							  where roleIds.Contains(menu.RoleId)
 							  select menu)
-							  .Select(m => m.NavigationMenu).Distinct().ToListAsync();
+							  .Select(m => m.NavigationMenu)
+							  .Distinct()
+							  .ToListAsync();
 
 			foreach (var item in data)
 			{
 				result = (item.ControllerName == ctrl && item.ActionName == act);
 				if (result)
+				{
 					break;
+				}
 			}
 
 			return result;
 		}
 
-		public async Task<List<NavigationMenuViewModel>> GetPermissionsByRoleIdAsync(string id)
+		public async Task<List<NavigationMenuViewModel>> GetPermissionsByRoleIdAsync(string? id)
 		{
 			var items = await (from m in _context.NavigationMenu
 							   join rm in _context.RoleMenuPermission
@@ -91,8 +104,10 @@ namespace Mvc.RoleAuthorization.Services
 			return items;
 		}
 
-		public async Task<bool> SetPermissionsByRoleIdAsync(string id, IEnumerable<Guid> permissionIds)
+		public async Task<bool> SetPermissionsByRoleIdAsync(string? id, IEnumerable<Guid> permissionIds)
 		{
+			if (string.IsNullOrWhiteSpace(id)) return false;
+
 			var existing = await _context.RoleMenuPermission.Where(x => x.RoleId == id).ToListAsync();
 			_context.RemoveRange(existing);
 
@@ -107,6 +122,9 @@ namespace Mvc.RoleAuthorization.Services
 
 			var result = await _context.SaveChangesAsync();
 
+			// Remove existing permissions to roles so it can re evaluate and take effect
+			_cache.Remove("RolePermissions");
+
 			return result > 0;
 		}
 
@@ -120,9 +138,7 @@ namespace Mvc.RoleAuthorization.Services
 			return data;
 		}
 
-		private string GetUserId(ClaimsPrincipal user)
-		{
-			return ((ClaimsIdentity)user.Identity).FindFirst(ClaimTypes.NameIdentifier)?.Value;
-		}
+		private static string? GetUserId(ClaimsPrincipal user) =>
+			(user.Identity) == null ? string.Empty : ((ClaimsIdentity)user.Identity).FindFirst(ClaimTypes.NameIdentifier)?.Value;
 	}
 }
